Change parameter type of obtenerPorDni method

+++ src/main/java/com/ProyectoDeAula5/Proyecto5/service/ClienteService.java
@@ -33,7 +33,7 @@ public class ClienteService {
         clienteRepository.deleteById(id);
     }
 
-    public Cliente obtenerPorDni(int dni) {
+    public Cliente obtenerPorDni(Long dni) {
         return clienteRepository.findByDni(dni).orElse(null);
     }
